moved from abstract-node to node
