+++ src/core/index.ts
@@ -1,4 +1,4 @@
-export * from "./nodes/abstract-node"
+export * from "./node"
 export * from "./action"
 export * from "./json-patch"
 export * from "./mst-operations"
