+++ src/index.ts
@@ -1,5 +1,5 @@
 // Fix some circular deps:
-import "./core/nodes/abstract-node"
+import "./core/node"
 import "./types/type"
 import "./types/complex-types/complex-type"
